Update html passages view
